# packages/api-gateway/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mentor Matching Platform API
  description: |
    A comprehensive API for the mentor matching platform that connects mentors and students through session-based mentoring.
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate limited:
    - General endpoints: 100 requests per 15 minutes
    - Authentication endpoints: 5 requests per 15 minutes
    
    ## Error Handling
    All errors follow a consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human readable error message"
      },
      "timestamp": "2024-01-01T00:00:00.000Z"
    }
    ```
  version: 1.0.0
  contact:
    name: Mentor Platform API Support
    email: api-support@mentorplatform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.mentorplatform.com/api
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for external integrations
  schemas:
    Error:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
    SuccessResponse:
      type: object
      required:
        - success
        - data
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data (varies by endpoint)
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T00:00:00.000Z"
...
    UserProfile:
      type: object
      required:
        - userId
        - firstName
        - lastName
      properties:
        userId:
          type: string
          format: uuid
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        bio:
          type: string
          example: "Experienced software engineer with 10+ years in web development"
        profileImageUrl:
          type: string
          format: uri
          example: "https://example.com/images/profile.jpg"
    MentorProfile:
      allOf:
        - $ref: '#/components/schemas/UserProfile'
        - type: object
          properties:
            expertiseAreas:
              type: array
              items:
                type: string
              example: ["JavaScript", "React", "Node.js"]
            yearsOfExperience:
              type: integer
              minimum: 0
              example: 10
            hourlyRate:
              type: number
              minimum: 0
              example: 75.00
            rating:
              type: number
              minimum: 0
              maximum: 5
              example: 4.8
            totalSessions:
              type: integer
              minimum: 0
              example: 150
    StudentProfile:
      allOf:
        - $ref: '#/components/schemas/UserProfile'
        - type: object
          properties:
            learningGoals:
              type: array
              items:
                type: string
              example: ["Learn React", "Improve JavaScript skills"]
            interests:
              type: array
              items:
                type: string
              example: ["Web Development", "Mobile Apps"]
            currentLevel:
              type: string
              enum: [BEGINNER, INTERMEDIATE, ADVANCED]
              example: "INTERMEDIATE"
    Session:
      type: object
      required:
        - id
        - mentorId
        - title
        - scheduledAt
        - duration
        - capacity
        - status
      properties:
        id:
          type: string
          format: uuid
        mentorId:
          type: string
          format: uuid
        title:
          type: string
          example: "Introduction to React Hooks"
        description:
          type: string
          example: "Learn the basics of React Hooks and how to use them effectively"
        expertiseAreas:
          type: array
          items:
            type: string
          example: ["React", "JavaScript"]
        sessionType:
          type: string
          enum: [ONE_ON_ONE, GROUP, WORKSHOP]
          example: "GROUP"
        scheduledAt:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in minutes
          example: 60
        capacity:
          type: integer
          minimum: 1
          example: 10
        currentRegistrations:
          type: integer
          minimum: 0
          example: 7
        status:
          type: string
          enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
          example: "SCHEDULED"
        meetingLink:
          type: string
          format: uri
          example: "https://meet.example.com/session-123"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
            timestamp: "2024-01-01T00:00:00.000Z"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "FORBIDDEN"
              message: "Insufficient permissions"
            timestamp: "2024-01-01T00:00:00.000Z"
...
    ValidationError:
      description: Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid input data"
            timestamp: "2024-01-01T00:00:00.000Z"
...
paths:
...
  /users/profile:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/MentorProfile'
                          - $ref: '#/components/schemas/StudentProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      tags:
        - Users
      summary: Update user profile
      description: Update the authenticated user's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                bio:
                  type: string
                expertiseAreas:
                  type: array
                  items:
                    type: string
                learningGoals:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Session endpoints
  /sessions:
    get:
      tags:
        - Sessions
      summary: Get available sessions
      description: Retrieve a list of available sessions with optional filtering
      parameters:
        - name: expertiseArea
          in: query
          schema:
            type: string
          description: Filter by expertise area
        - name: sessionType
          in: query
          schema:
            type: string
            enum: [ONE_ON_ONE, GROUP, WORKSHOP]
          description: Filter by session type
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of sessions to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of sessions to skip
      responses:
        '200':
          description: Sessions retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          sessions:
                            type: array
                            items:
                              $ref: '#/components/schemas/Session'
                          total:
                            type: integer
                          limit:
                            type: integer
                          offset:
                            type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Sessions
      summary: Create a new session
      description: Create a new mentoring session (mentors only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - scheduledAt
                - duration
                - capacity
                - sessionType
              properties:
                title:
                  type: string
                description:
                  type: string
                expertiseAreas:
                  type: array
                  items:
                    type: string
                sessionType:
                  type: string
                  enum: [ONE_ON_ONE, GROUP, WORKSHOP]
                scheduledAt:
                  type: string
                  format: date-time
                duration:
                  type: integer
                  minimum: 15
                  maximum: 480
                capacity:
                  type: integer
                  minimum: 1
                  maximum: 100
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  # Webhook endpoints
...